move nunique, trunctate
